Added PostMapping and GetAll services H2 Policy schema and data... testing... v3

+++ src/main/resources/schema-h2.sql
@@ -1,7 +1,7 @@
 /*
 create table policy
 (
-    policyNo integer not null,
+    policyNo bigint not null,
     firstName varchar(50) not null,
     lastName varchar(50) not null,
     startDate date not null,
